requirements.txt
# ─── Web ───
python-dotenv==1.0.1
streamlit

# ─── 資料庫 ───
sqlalchemy==2.0.28
psycopg2-binary==2.9.9

# ─── 股票數據 ───
yfinance==1.2.0
pandas==2.2.1
numpy==1.26.4

# ─── 回測引擎 ───
pandas-ta @ git+https://github.com/aarigs/pandas-ta
vectorbt==0.26.0
numba==0.59.0
quantstats==0.0.81

# ─── AI / 機器學習 ───
groq==0.5.0
xgboost==3.2.0
scikit-learn==1.8.0

# ─── 圖表 ───
plotly==5.19.0
matplotlib==3.10.8

# ─── 通知 ───
line-bot-sdk==3.22.0

# ─── 排程 ───
apscheduler==3.11.2
pytz==2026.1.post1

# ─── 工具 ───
beautifulsoup4==4.14.3